Refactor joins in `self.hash_tree`
Eliminate the need for the self cross join to get 
root ids, a simpler query execution plan, and a 
little more scope re-use. This should lessen the
impact of calling hash_tree for very large 
datasets.

Still need to join to the grouped hierarchy table
to get generation depth but use it as a simple
inner join with `descendant_id` and the max 
`generations` filtered by the `limit_depth`.